refatoração do componente de modal
