+++ src/components/screens/Logistica/ConsultaEntregas/components/ListagemSolicitacoes/index.jsx
@@ -1,5 +1,4 @@
 import React, { useState, useContext } from "react";
-import { Modal } from "react-bootstrap";
 import "antd/dist/antd.css";
 import "./styles.scss";
 import FiltrosExcel from "../FiltrosExcel";
@@ -7,6 +6,7 @@ import { imprimirGuiasDaSolicitacao } from "services/logistica.service.js";
 import { toastError } from "components/Shareable/Toast/dialogs";
 import { Spin } from "antd";
 import { CentralDeDownloadContext } from "context/CentralDeDownloads";
+import ModalSolicitacaoDownload from "components/Shareable/ModalSolicitacaoDownload";
 
 const ListagemSolicitacoes = ({ solicitacoes, ativos, setAtivos, dilog }) => {
   const [carregando, setCarregando] = useState(false);
@@ -28,21 +28,9 @@ const ListagemSolicitacoes = ({ solicitacoes, ativos, setAtivos, dilog }) => {
       });
   };
 
-  const handleClose = () => {
-    setShow(false);
-  };
-
   return (
     <Spin tip="Carregando..." spinning={carregando}>
-      <Modal show={show} onHide={handleClose} dialogClassName="modal-entregas">
-        <Modal.Header closeButton>
-          <Modal.Title>Geração solicitada com sucesso.</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          A geração foi solicitada. Em breve você receberá um aviso na central
-          de downloads com o resultado.
-        </Modal.Body>
-      </Modal>
+      <ModalSolicitacaoDownload show={show} setShow={setShow} />
       <section
         className={`resultado-busca-entregas ${
           dilog ? "dilog" : "distribuidor"
